requirements.txt
boto3==1.18.33
fastapi==0.68.1
Pillow==8.3.1
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
uvicorn==0.15.0
